10 Спринт, Доработка, Доавлен метод isNotExist 0.3
